viSer-3835 - Creation of a common library for all applications
- moved dialog popups from Activity to Context
