+++ app/src/main/java/com/smartviser/androidext/ContextExt.kt
@@ -2,7 +2,9 @@
 
 package com.smartviser.androidext
 
+import android.app.ProgressDialog
 import android.content.Context
+import android.content.DialogInterface
 import android.content.pm.PackageManager
 import android.os.Build
 import android.telephony.SubscriptionManager
@@ -10,6 +12,8 @@ import android.telephony.TelephonyManager
 import androidx.annotation.RequiresApi
 import android.content.res.Resources
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
 
 val Context.telephonyManager: TelephonyManager
     get() = getSystemService(Context.TELEPHONY_SERVICE) as TelephonyManager
@@ -23,8 +27,8 @@ fun Context.checkPermission(vararg permissions: String): Boolean =
         accumulator && checkSelfPermission(permission) == PackageManager.PERMISSION_GRANTED
     }
 
-fun Context.toast(message: String, isShort: Boolean) =
-    Toast.makeText(this, message, if (isShort) Toast.LENGTH_SHORT else Toast.LENGTH_LONG).show()
+// Resources
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 fun Context.getResourcesArrayValue(arrayId: Int, position: Int): String? =
     try {
@@ -32,3 +36,31 @@ fun Context.getResourcesArrayValue(arrayId: Int, position: Int): String? =
     } catch (e: Resources.NotFoundException) {
         null
     }
+
+// Popups
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+fun AppCompatActivity.popup(titleId: Int?, messageId: Int) {
+    popup(titleId, messageId, null)
+}
+
+fun AppCompatActivity.popup(
+    titleId: Int?,
+    messageId: Int,
+    listener: DialogInterface.OnClickListener?
+) {
+    val builder = AlertDialog.Builder(this)
+    if (titleId != null) {
+        builder.setTitle(titleId)
+    }
+    builder.setCancelable(false)
+    builder.setMessage(messageId)
+    builder.setPositiveButton("OK", listener)
+    builder.create().show()
+}
+
+fun AppCompatActivity.waitPopup(): ProgressDialog =
+    ProgressDialog.show(this, "Loading", "Please wait...", true)
+
+fun Context.toast(message: String, isShort: Boolean) =
+    Toast.makeText(this, message, if (isShort) Toast.LENGTH_SHORT else Toast.LENGTH_LONG).show()
